% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{characteristic_names_lookup}
\alias{characteristic_names_lookup}
\title{DCR Characteristic Names}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7 rows and 12 columns.
}
\usage{
characteristic_names_lookup
}
\description{
these are the characteristics collected and reported by DCR
}
\keyword{datasets}
