% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_append_data}
\alias{db_append_data}
\title{Append Data}
\usage{
db_append_data(db_connection, table_name, data, overlaps = FALSE)
}
\arguments{
\item{db_connection}{a connection the local database}

\item{table_name}{name of the table to append to}

\item{data}{data to add}
}
\description{
add data to an existing table on the database
}
