% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fix_characteristic_name}
\alias{fix_characteristic_name}
\title{Fix Characteristic Names}
\usage{
fix_characteristic_name(name)
}
\description{
when needed it transforms the characteristic name as reported from the lab
to the WQX compliant version
}
