% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-transducers.R
\name{parse_transducer_stage}
\alias{parse_transducer_stage}
\title{Structure Transducer for WQX Stage}
\usage{
parse_transducer_stage(data, metadata)
}
\arguments{
\item{data}{Transducer data ready to be restructured to WQX}
}
\description{
Structure Transducer for WQX Stage
}
