% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-transducers.R
\name{transducer_col_info}
\alias{transducer_col_info}
\title{Column info needed to read data}
\usage{
transducer_col_info(serial)
}
\description{
return a list with the column headers and column
types needed to read in in the data using read_csv
}
