% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-transducers.R
\name{transducer_properties}
\alias{transducer_properties}
\title{Transducer Properties}
\usage{
transducer_properties(v)
}
\arguments{
\item{v}{a vector whose elements are each line in the text file}
}
\description{
Get transducer device properties from the text file
}
