% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transducers}
\alias{transducers}
\title{Dry Creek Rancheria Active Transducers}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 6 columns.
}
\usage{
transducers
}
\description{
these are transducers currently in operation
}
\keyword{datasets}
