% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqx.R
\name{upload_data}
\alias{upload_data}
\title{Upload Data to WQX}
\usage{
upload_data(file, type, config, username, pk)
}
\arguments{
\item{file}{path to file to upload}

\item{type}{what type of file is this}

\item{config}{path to config file}

\item{username}{username}

\item{pk}{private key}
}
\description{
function uploads
}
