% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-transducers.R
\name{read_transducer}
\alias{read_transducer}
\title{Read transducer data}
\usage{
read_transducer(file)
}
\arguments{
\item{file}{a csv file with transducer output}
}
\description{
read transducer raw data and restructure to conform
with DCR and WQX compliant schema
}
