% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-transducers.R
\name{transducer_col_info}
\alias{transducer_col_info}
\title{Transducer Column Info}
\usage{
transducer_col_info(serial)
}
\arguments{
\item{serial}{the serial number for the transducer}
}
\description{
based on a transducer's serial number determine
the column names and column needed to read in the csv
}
